Remove non-InCommon entities from InCommon aggregate.
This allows an existing InCommon production aggregate (including
entities from eduGAIN) to be used as a source for the eduGAIN import
flow. Without this, the entities already in the aggregate override the
newly imported ones from eduGAIN.

This should not have any effect in production, where the InCommon
source should just be InCommon entities anyway. It is most valuable
while testing the eduGAIN import flow.